requirements.txt
flask
python-dotenv
markdown
flask_sqlalchemy
requests
google-generativeai
Flask-Cors 